Fixed Clang compilation warnings;

+++ unittest/shell_command_handler_t.cc
@@ -86,7 +86,7 @@ namespace shcore {
 
       EXPECT_TRUE(_shell_command_handler.process("cmdone parameter"));
       EXPECT_EQ("cmd_one", _function);
-      ASSERT_EQ(1, _params.size());
+      ASSERT_EQ(1, static_cast<int>(static_cast<int>(_params.size())));
       EXPECT_EQ("parameter", _params[0]);
 
       EXPECT_TRUE(_shell_command_handler.process("cmd2"));
@@ -95,7 +95,7 @@ namespace shcore {
 
       EXPECT_TRUE(_shell_command_handler.process("\\2 two parameters"));
       EXPECT_EQ("cmd_two", _function);
-      ASSERT_EQ(2, _params.size());
+      ASSERT_EQ(2, static_cast<int>(static_cast<int>(_params.size())));
       EXPECT_EQ("two", _params[0]);
       EXPECT_EQ("parameters", _params[1]);
 
@@ -105,7 +105,7 @@ namespace shcore {
 
       EXPECT_TRUE(_shell_command_handler.process("three three different parameters"));
       EXPECT_EQ("cmd_three", _function);
-      ASSERT_EQ(3, _params.size());
+      ASSERT_EQ(3, static_cast<int>(_params.size()));
       EXPECT_EQ("three", _params[0]);
       EXPECT_EQ("different", _params[1]);
       EXPECT_EQ("parameters", _params[2]);
@@ -116,7 +116,7 @@ namespace shcore {
 
       EXPECT_TRUE(_shell_command_handler.process("\\4 now four different parameters"));
       EXPECT_EQ("cmd_four", _function);
-      ASSERT_EQ(4, _params.size());
+      ASSERT_EQ(4, static_cast<int>(_params.size()));
       EXPECT_EQ("now", _params[0]);
       EXPECT_EQ("four", _params[1]);
       EXPECT_EQ("different", _params[2]);
